Changes to read_cases.
Moving forward in fads a few years, read_cases now takes in a JSON
file instead of an SQLite3 database. This has the advantage of making
it human-readable in raw form and thus friendlier.

The aliases dictionary is now keyed on (court, number) instead of only
number, since it is uncertain whether or not PACER numbers are
unique across courts (probably not).